LayoutTests/http/tests/security/contentSecurityPolicy/report-uri-expected.txt

CONSOLE MESSAGE: Refused to execute inline script because of Content-Security-Policy.

CSP report received:
CONTENT_TYPE: application/x-www-form-urlencoded
HTTP_REFERER: http://127.0.0.1:8000/security/contentSecurityPolicy/report-uri.html
REQUEST_METHOD: POST
=== POST DATA ===
document-url: http://127.0.0.1:8000/security/contentSecurityPolicy/report-uri.html
violated-directive: script-src 'self'
